refactor: reorganize frontend components into domain folders
Groups loose root-level components into events/, notifications/, map/,
profile/, settings/, and center/ subfolders. Eliminates the web/ folder
(files renamed to .web.tsx and moved to domain folders). Moves storage
utilities (tokenStorage, onboardingStorage) out of components/utils/ into
src/storage/ and utils/ respectively. Moves connect/mockData.ts into
__mocks__/ subfolder.
