+++ packages/frontend/app/(tabs)/_layout.tsx
@@ -4,7 +4,7 @@ import { useState, useEffect } from 'react'
 import { useUser, useTheme } from '../../components/contexts'
 import { HeaderActionProvider } from '../../components/contexts/HeaderActionContext'
 import { House, MessageSquare, Newspaper, Compass, Plus, Bell } from 'lucide-react-native'
-import SettingsPanel from '../../components/SettingsPanel'
+import SettingsPanel from '../../components/settings/SettingsPanel'
 import Logo from '../../components/ui/Logo'
 import TabHeader from '../../components/ui/TabHeader'
 import { Avatar } from '../../components/ui'
