+++ packages/frontend/components/auth/PasswordStrength.tsx
