+++ packages/frontend/components/center/CenterDetailPanel.web.tsx
@@ -6,8 +6,8 @@ import UnderlineTabBar from '../ui/UnderlineTabBar'
 import type { CenterDisplay } from '../../hooks/useApiData'
 import type { EventDisplay } from '../../utils/api'
 import { useDetailColors } from '../../hooks/useDetailColors'
-import { buildCenterBoard, ThreadPanel } from '../../components/connect'
-import { useUser } from '../../components/contexts'
+import { buildCenterBoard, ThreadPanel } from '../connect'
+import { useUser } from '../contexts'
 
 // ── Props ────────────────────────────────────────────────────────────────
