+++ packages/frontend/components/connect/index.ts
@@ -1,3 +1,3 @@
 export { BoardPostCard, ThreadPanel } from './ThreadPanel'
 export type { ComposerState, ThreadPanelColors } from './ThreadPanel'
-export * from './mockData'
+export * from './__mocks__/mockData'
