+++ packages/frontend/components/home/HomeMiniMap.web.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Pressable, Text, View } from 'react-native'
 import { Map as MapIcon } from 'lucide-react-native'
-import CommunityMap from '../Map'
+import CommunityMap from '../map/Map'
 import type { MapPoint } from '../../utils/api'
 
 type TilePreview = {
