+++ packages/frontend/components/notifications/NotificationItem.tsx
@@ -6,8 +6,8 @@
 
 import React from 'react'
 import { View, Text, TouchableOpacity, StyleSheet } from 'react-native'
-import type { Notification } from '../utils/notificationService'
-import { getNotificationTypeName } from '../utils/notificationService'
+import type { Notification } from '../../utils/notificationService'
+import { getNotificationTypeName } from '../../utils/notificationService'
 
 interface NotificationItemProps {
   notification: Notification
