+++ packages/frontend/components/profile/AvatarCropper.tsx
