+++ packages/frontend/components/profile/AvatarCropper.web.tsx
